The update_ranges.py is now the new mv method with the old method available as update_ranges_old.py. update_ranges now has backwards compatible option handling (with warnings).
