--- datacube_ows/product_ranges_2.py
@@ -1,351 +1,0 @@
-#pylint: skip-file
-
-from __future__ import absolute_import, division, print_function
-
-from datetime import datetime
-import datacube
-
-from datacube_ows.ows_configuration import get_config, OWSNamedLayer  # , get_layers, ProductLayerDef
-from datacube_ows.ogc_utils import local_date
-from psycopg2.extras import Json
-from itertools import zip_longest
-import json
-
-from datacube_ows.utils import get_sqlconn
-
-def get_odc_products(dc, any_product, odc_only=False):
-    if isinstance(any_product, OWSNamedLayer):
-        return any_product.products
-    elif isinstance(any_product, str):
-        dc_product = dc.index.products.get_by_name(any_product)
-        if odc_only:
-            ows_product = None
-        else:
-            ows_product = get_config().product_index.get(any_product)
-        if ows_product:
-            if dc_product and [dc_product] == ows_product.products:
-                # The same!
-                return [dc_product]
-            print("Updating OWS product %s (ODC Products: []).  If you meant the ODC product %s, please use the --odc-only flag.")
-            return ows_product.products
-    else:
-        # Assume ODC product
-        return [any_product]
-
-def get_crsids(cfg=None):
-    if not cfg:
-        cfg = get_config()
-    return cfg.published_CRSs.keys()
-
-
-def get_crses(cfg=None):
-    return  {crsid: datacube.utils.geometry.CRS(crsid) for crsid in get_crsids(cfg)}
-
-
-def jsonise_bbox(bbox):
-    if isinstance(bbox, dict):
-        return bbox
-    else:
-        return {
-            "top": bbox.top,
-            "bottom": bbox.bottom,
-            "left": bbox.left,
-            "right": bbox.right,
-        }
-
-
-def create_multiprod_range_entry(dc, product, crses):
-    print("Merging multiproduct ranges for %s (ODC products: %s)" % (
-        product.name,
-        repr(product.product_names)
-    ))
-    conn = get_sqlconn(dc)
-    txn = conn.begin()
-    prodids = [ p.id for p in product.products ]
-    wms_name = product.name
-
-    # Attempt to insert row
-    conn.execute("""
-        INSERT INTO wms.multiproduct_ranges
-        (wms_product_name,lat_min,lat_max,lon_min,lon_max,dates,bboxes)
-        VALUES
-        (%(p_id)s, 0, 0, 0, 0, %(empty)s, %(empty)s)
-        ON CONFLICT (wms_product_name) DO NOTHING
-        """,
-                 {"p_id": wms_name, "empty": Json("")})
-
-    # Update extents
-    conn.execute("""
-        UPDATE wms.multiproduct_ranges
-        SET lat_min = subq.lat_min,
-            lat_max = subq.lat_max, 
-            lon_min = subq.lon_min, 
-            lon_max = subq.lon_max
-        FROM (
-            select min(lat_min) as lat_min,
-                   max(lat_max) as lat_max,
-                   min(lon_min) as lon_min,
-                   max(lon_max) as lon_max
-            from wms.product_ranges
-            where id = ANY (%(p_prodids)s)
-        ) as subq
-        WHERE wms_product_name = %(p_id)s
-        """,
-                 {"p_id": wms_name, "p_prodids": prodids})
-
-    # Create sorted list of dates
-    results = conn.execute(
-        """
-        SELECT dates
-        FROM   wms.product_ranges
-        WHERE  id  = ANY (%(p_prodids)s)
-        """, {"p_prodids": prodids}
-    )
-    dates = set()
-    for r in results:
-        for d in r[0]:
-            dates.add(d)
-    dates = sorted(dates)
-    conn.execute("""
-           UPDATE wms.multiproduct_ranges
-           SET dates = %(dates)s
-           WHERE wms_product_name= %(p_id)s
-      """,
-                 {
-                     "dates": Json(dates),
-                     "p_id": wms_name
-                 }
-    )
-
-    # calculate bounding boxes
-    results = list(conn.execute("""
-        SELECT lat_min,lat_max,lon_min,lon_max
-        FROM wms.multiproduct_ranges
-        WHERE wms_product_name=%(p_id)s
-        """,
-        {"p_id": wms_name} ))
-
-    r = results[0]
-
-    epsg4326 = datacube.utils.geometry.CRS("EPSG:4326")
-    box = datacube.utils.geometry.box(
-        float(r[2]),
-        float(r[0]),
-        float(r[3]),
-        float(r[1]),
-        epsg4326)
-
-    cfg = get_config()
-    conn.execute("""
-        UPDATE wms.multiproduct_ranges
-        SET bboxes = %s::jsonb
-        WHERE wms_product_name=%s
-        """,
-                 Json({ crsid: jsonise_bbox(box.to_crs(crs).boundingbox) for crsid, crs in get_crses(cfg).items() }),
-                 wms_name
-    )
-
-    txn.commit()
-    conn.close()
-    return
-
-
-def create_range_entry(dc, product, crses, summary_product=False):
-  print("Updating range for ODC product %s..."% product.name)
-  # NB. product is an ODC product
-  conn = get_sqlconn(dc)
-  txn = conn.begin()
-  prodid = product.id
-
-  # insert empty row if one does not already exist
-  conn.execute("""
-    INSERT INTO wms.product_ranges
-    (id,lat_min,lat_max,lon_min,lon_max,dates,bboxes)
-    VALUES
-    (%(p_id)s, 0, 0, 0, 0, %(empty)s, %(empty)s)
-    ON CONFLICT (id) DO NOTHING
-    """,
-    {"p_id": prodid, "empty": Json("")})
-
-
-  # Update min/max lat/longs
-  conn.execute(
-      """
-      UPDATE wms.product_ranges pr
-      SET lat_min = st_ymin(subq.bbox),
-          lat_max = st_ymax(subq.bbox),
-          lon_min = st_xmin(subq.bbox),
-          lon_max = st_xmax(subq.bbox)
-      FROM (
-        SELECT st_extent(stv.spatial_extent) as bbox
-        FROM public.space_time_view stv
-        WHERE stv.dataset_type_ref = %(p_id)s
-      ) as subq
-      """,
-      {"p_id": prodid})
-
-  # Set default timezone
-  conn.execute("""
-    set timezone to 'Etc/UTC'
-  """)
-
-
-  if summary_product:
-      # Loop over dates
-      dates = set()
-
-      results = conn.execute(
-          """
-          select  
-                array_agg(temporal_extent)
-          from public.space_time_view 
-          WHERE dataset_type_ref = %(p_id)s
-          """,
-          {"p_id": prodid}
-      )
-      for result in results:
-          for dat_ran in result[0]:
-              dates.add(dat_ran.lower)
-
-      dates = sorted(dates)
-
-      conn.execute("""
-           UPDATE wms.product_ranges
-           SET dates = %(dates)s
-           WHERE id= %(p_id)s
-      """,
-                   {
-                       "dates": Json([t.strftime("%Y-%m-%d") for t in dates]),
-                       "p_id": prodid
-                   }
-      )
-  else:
-      # Create sorted list of dates
-      conn.execute("""
-        WITH sorted
-        AS (SELECT to_jsonb(array_agg(dates.d))
-            AS dates
-            FROM (SELECT DISTINCT
-                   date(cast(metadata -> 'extent' ->> 'center_dt' as timestamp) AT TIME ZONE 'UTC' +
-                    (least(to_number(metadata -> 'extent' -> 'coord' -> 'll' ->> 'lon', '9999.9999999999999999999999999999999999'),
-                    to_number(metadata -> 'extent' -> 'coord' -> 'ul' ->> 'lon', '9999.9999999999999999999999999999999999')) +
-                    greatest(to_number(metadata -> 'extent' -> 'coord' -> 'lr' ->> 'lon', '9999.9999999999999999999999999999999999'),
-                    to_number(metadata -> 'extent' -> 'coord' -> 'ur' ->> 'lon', '9999.9999999999999999999999999999999999'))) / 30.0 * interval '1 hour')
-                  AS d
-                  FROM agdc.dataset
-                  WHERE dataset_type_ref=%(p_id)s
-                  AND archived IS NULL
-                  ORDER BY d) dates)
-        UPDATE wms.product_ranges
-        SET dates=sorted.dates
-        FROM sorted
-        WHERE id=%(p_id)s
-        """,
-        {"p_id": prodid})
-
-
-  # calculate bounding boxes
-  results = list(conn.execute("""
-    SELECT lat_min,lat_max,lon_min,lon_max
-    FROM wms.product_ranges
-    WHERE id=%s
-    """,
-    prodid))
-
-  r = results[0]
-
-  epsg4326 = datacube.utils.geometry.CRS("EPSG:4326")
-  box = datacube.utils.geometry.box(
-    float(r[2]),
-    float(r[0]),
-    float(r[3]),
-    float(r[1]),
-    epsg4326)
-
-  conn.execute("""
-    UPDATE wms.product_ranges
-    SET bboxes = %s::jsonb
-    WHERE id=%s
-    """,
-    Json(
-      {crsid: {"top": box.to_crs(crs).boundingbox.top,
-               "bottom": box.to_crs(crs).boundingbox.bottom,
-               "left": box.to_crs(crs).boundingbox.left,
-               "right": box.to_crs(crs).boundingbox.right}
-        for crsid, crs in crses.items()
-       }
-    ),
-    product.id)
-
-  txn.commit()
-  conn.close()
-
-
-def check_datasets_exist(dc, product_name):
-  conn = get_sqlconn(dc)
-
-  results = conn.execute("""
-    SELECT COUNT(*)
-    FROM agdc.dataset ds, agdc.dataset_type p
-    WHERE ds.archived IS NULL
-    AND ds.dataset_type_ref = p.id
-    AND p.name = %s""",
-    product_name)
-
-  conn.close()
-
-  return list(results)[0][0] > 0
-
-def add_ranges(dc, product_names, summary=False, merge_only=False):
-    odc_products = {}
-    ows_multiproducts = []
-    for pname in product_names:
-        dc_product = None
-        ows_product = get_config().product_index.get(pname)
-        if not ows_product:
-            ows_product = get_config().native_product_index.get(pname)
-        if ows_product:
-            for dc_pname in ows_product.product_names:
-                if dc_pname in odc_products:
-                    odc_products[dc_pname]["ows"].append(ows_product)
-                else:
-                    odc_products[dc_pname] = { "ows": [ows_product]}
-            print("OWS Layer %s maps to ODC Product(s): %s" % (
-                ows_product.name,
-                repr(ows_product.product_names)
-            ))
-            if ows_product.multi_product:
-                ows_multiproducts.append(ows_product)
-        if not ows_product:
-            dc_product = dc.index.products.get_by_name(pname)
-            if dc_product:
-                print("ODC Layer: %s" % pname)
-                if pname in odc_products:
-                    odc_products[pname]["ows"].append(None)
-                else:
-                    odc_products[pname] = { "ows": [None]}
-            else:
-                print("Unrecognised product name:", pname)
-                continue
-
-    if ows_multiproducts and merge_only:
-        print("Merge-only: Skipping range update of products:", repr(list(odc_products.keys())))
-    else:
-        for pname, ows_prods in odc_products.items():
-            dc_product = dc.index.products.get_by_name(pname)
-            if check_datasets_exist(dc, dc_product.name):
-                prod_summary = summary
-                for ows_prod in ows_prods["ows"]:
-                    if ows_prod:
-                        prod_summary = not ows_prod.is_raw_time_res
-                        break
-                create_range_entry(dc, dc_product, get_crses(), prod_summary)
-            else:
-                print("Could not find any datasets for: ", pname)
-
-    for mp in ows_multiproducts:
-        create_multiprod_range_entry(dc, mp, get_crses())
-
-    print("Done.")
-
-
