--- update_ranges_2.py
@@ -1,4 +1,0 @@
-from datacube_ows.update_ranges_2 import main
-
-if __name__ == '__main__':
-    main()
