+++ update_ranges_old.py
@@ -1,0 +1,4 @@
+from datacube_ows.update_ranges_old import main
+
+if __name__ == '__main__':
+    main()
